+++ src/fci/fci_integrals.cc
@@ -188,7 +188,7 @@ void FCIIntegrals::set_active_integrals_and_restricted_docc() {
     RestrictedOneBodyOperator(oei_a_, oei_b_);
 }
 
-double FCIIntegrals::energy(const STLDeterminant& det) {
+double FCIIntegrals::energy(const STLBitsetDeterminant& det) {
     const bit_t bits = det.bits();
     double energy = frozen_core_energy_;
     for (int p = 0; p < nmo_; p++) {
@@ -217,7 +217,7 @@ double FCIIntegrals::energy(const STLDeterminant& det) {
     return energy;
 }
 
-double FCIIntegrals::energy(STLDeterminant& det) {
+double FCIIntegrals::energy(STLBitsetDeterminant& det) {
     const bit_t bits = det.bits();
     double energy = frozen_core_energy_;
     for (int p = 0; p < nmo_; p++) {
@@ -246,17 +246,17 @@ double FCIIntegrals::energy(STLDeterminant& det) {
     return energy;
 }
 
-STLDeterminant FCIIntegrals::determinant() {
-    STLDeterminant det(nmo_);
+STLBitsetDeterminant FCIIntegrals::determinant() {
+    STLBitsetDeterminant det(nmo_);
     return det;
 }
 
-STLDeterminant FCIIntegrals::determinant(const bit_t& bits) {
-    STLDeterminant det(bits, nmo_);
+STLBitsetDeterminant FCIIntegrals::determinant(const bit_t& bits) {
+    STLBitsetDeterminant det(bits, nmo_);
     return det;
 }
 
-double FCIIntegrals::slater_rules(const STLDeterminant& lhs, const STLDeterminant& rhs) const {
+double FCIIntegrals::slater_rules(const STLBitsetDeterminant& lhs, const STLBitsetDeterminant& rhs) const {
     const bit_t& I = lhs.bits_;
     const bit_t& J = rhs.bits_;
 
@@ -426,7 +426,7 @@ double FCIIntegrals::slater_rules(const STLDeterminant& lhs, const STLDeterminan
     return (matrix_element);
 }
 
-double FCIIntegrals::slater_rules_single_alpha(const STLDeterminant& det, int i, int a) const {
+double FCIIntegrals::slater_rules_single_alpha(const STLBitsetDeterminant& det, int i, int a) const {
     // Slater rule 2 PhiI = j_a^+ i_a PhiJ
     const bit_t& I = det.bits();
     double sign = det.slater_sign_aa(i, a);
@@ -442,7 +442,7 @@ double FCIIntegrals::slater_rules_single_alpha(const STLDeterminant& det, int i,
     return sign * matrix_element;
 }
 
-double FCIIntegrals::slater_rules_single_alpha_abs(const STLDeterminant& det, int i, int a) const {
+double FCIIntegrals::slater_rules_single_alpha_abs(const STLBitsetDeterminant& det, int i, int a) const {
     // Slater rule 2 PhiI = j_a^+ i_a PhiJ
     const bit_t& I = det.bits();
     double matrix_element = oei_a_[i * nmo_ + a];
@@ -457,7 +457,7 @@ double FCIIntegrals::slater_rules_single_alpha_abs(const STLDeterminant& det, in
     return matrix_element;
 }
 
-double FCIIntegrals::slater_rules_single_beta(const STLDeterminant& det, int i, int a) const {
+double FCIIntegrals::slater_rules_single_beta(const STLBitsetDeterminant& det, int i, int a) const {
     // Slater rule 2 PhiI = j_a^+ i_a PhiJ
     const bit_t& I = det.bits();
     double sign = det.slater_sign_bb(i, a);
@@ -473,7 +473,7 @@ double FCIIntegrals::slater_rules_single_beta(const STLDeterminant& det, int i, 
     return sign * matrix_element;
 }
 
-double FCIIntegrals::slater_rules_single_beta_abs(const STLDeterminant& det, int i, int a) const {
+double FCIIntegrals::slater_rules_single_beta_abs(const STLBitsetDeterminant& det, int i, int a) const {
     // Slater rule 2 PhiI = j_a^+ i_a PhiJ
     const bit_t& I = det.bits();
     double matrix_element = oei_b_[i * nmo_ + a];
