+++ src/fci_mo.cc
@@ -89,7 +89,6 @@ void FCI_MO::startup() {
     ambit::Tensor tei_active_bb = integral_->aptei_bb_block(idx_a_, idx_a_, idx_a_, idx_a_);
     fci_ints_->set_active_integrals(tei_active_aa, tei_active_ab, tei_active_bb);
     fci_ints_->compute_restricted_one_body_operator();
-    STLBitsetDeterminant::set_ints(fci_ints_);
 
     // compute orbital extents if CIS/CISD IPEA
     if (ipea_ != "NONE") {
